Relatively decent

+++ src/config.rs
@@ -43,7 +43,7 @@ def_cfg! {
         possible_settings: Option<Vec<CameraSettings>>,
         subsystems: Subsystems,
         calib: Option<serde_json::Value>,
-        gamma: Option<f32>,
+        gamma: Option<f64>,
     }
     CameraSettings {
         width: u32,
